Main Code Upload
This code is still a work in progress as I need to finalise the timer to countdown - as well as useful headless audio outputs for startup and debugging. The hardware and setup can be seen on my website www.elliotthall.co.uk when I have chance to upload the project.